requirements.txt
altgraph==0.17.4
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
distlib==0.3.8
filelock==3.13.1
idna==3.6
Nuitka==1.9.7
numpy==1.26.4
opencv-python==4.9.0.80
ordered-set==4.1.0
packaging==23.2
pefile==2023.2.7
pillow==10.2.0
pipenv==2023.11.17
platformdirs==4.1.0
py2exe==0.13.0.1
pycparser==2.21
pydub==0.25.1
pygame==2.5.2
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
pymunk==6.6.0
pywin32-ctypes==0.2.2
requests==2.31.0
SpeechRecognition==3.10.1
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0
zstandard==0.22.0
